--- a/vllm/v1/engine/core.py
+++ b/vllm/v1/engine/core.py
@@ -447,13 +447,6 @@ class EngineCore:
 
         req = Request.from_engine_core_request(request,
                                                self.request_block_hasher)
-        if req.use_structured_output:
-            # Note on thread safety: no race condition.
-            # `grammar_init` is only invoked in input processing thread. For
-            # `structured_output_manager`, each request is independent and
-            # grammar compilation is async. Scheduler always checks grammar
-            # compilation status before scheduling request.
-            self.structured_output_manager.grammar_init(req)
         return req, request.current_wave