--- a/vllm/v1/structured_output/__init__.py
+++ b/vllm/v1/structured_output/__init__.py
@@ -271,7 +271,7 @@ class StructuredOutputManager:
         return True
 
     def should_advance(self, request: CachedRequestState) -> bool:
-        if not request.sampling_params.guided_decoding:
+        if request.sampling_params.guided_decoding is None:
             return False
 
         # To determine whether we can advance the FSM.
@@ -288,7 +288,7 @@ class StructuredOutputManager:
                 return True
 
             # Check if reasoning ends in *this* step
-            all_token_ids = ConstantList(request.prompt_token_ids) + ConstantList(request.output_token_ids)
+            all_token_ids = ConstantList(request.prompt_token_ids + request.output_token_ids)
             if self.reasoner.is_reasoning_end(all_token_ids):
                 # Reasoning just ended, so we shouldn't advance til
                 # next pass