… benchmarks/benchmark_serving_structured_output.py
… vllm/v1/worker/tpu_model_runner.py