@njhill 2025-09-05
See also: https://github.com/vllm-project/vllm/pull/23224