@benchislett 2025-09-15
At a glance, I prefer the implementation in https://github.com/vllm-project/vllm/pull/23224 to this one, though they're very similar. I've messaged the author of that PR and am hoping we can make progress there.