@russellb 2025-10-30
Thank you for the PR!

There have been a few different efforts in this area. The one we're moving forward with right now is #26866. Any feedback you have on that appraoch is welcome!